Show a single summary statistic in a text span

A summaryNumberSpan displays a single statistic derived from a linked table. Its primary use is with the crosstalk package. Used with crosstalk, a summaryNumberSpan displays a value which updates as the data selection changes.

Usage

summaryNumberSpan(
  data,
  statistic = c("count", "sum", "mean", "pct_total", "sum_pct_total", "min", "max",
    "quantile", "wt_mean", "sum_ratio"),
  column = NULL,
  column2 = NULL,
  weight = NULL,
  selection = NULL,
  locale = "navigator.language",
  digits = 0,
  number_format = c("decimal", "percent", "currency", "unit"),
  signDisplay = c("auto", "always", "exceptZero", "negative", "never"),
  currency = NULL,
  unit = NULL,
  quantile = 0.5,
  notation = c("standard", "scientific", "engineering", "compact"),
  width = NULL,
  height = NULL,
  elementId = NULL
)

Arguments

data

Data to summarize, normally an instance of crosstalk::SharedData.

statistic

The statistic to compute. Possible vales are c("count", "sum", "mean", "pct_total", "sum_pct_total", "min", "max", "quantile", "wt_mean","sum_ratio"). Count is the default.

column

For most statistics, the column of data to summarize. Not used for "count" or "pct_total" statistic.

column2

Valid only when statistic = "sum_ratio". column2 is a column of data such that sum_ratio = sum(column)/sum(column2)

weight

Valid only when statistic = "wt_mean". Specifies the weight column of data.

selection

Expression to select a fixed subset of data. May be a logical vector or a one-sided formula that evaluates to a logical vector. If used, the key given to crosstalk::SharedData must be a fixed column (not row numbers).

locale

Control how the number is displayed. Valid strings include a IETF BCP 47 language tag, or the string "navigator.language". Default value is "navigator.language"

digits

Number of decimal places to display, or NULL to display full precision. if number_format == "percent", digits == 0 the statistics is 0.1234, the resulting visualization will show "12%".

number_format

The resulting type of number to display. Options are c("decimal", "percent", "currency","unit"). Decimal is the default.

signDisplay

How to display the sign for the number. Possible values are: c("auto", "always","exceptZero","negative","never"). Auto is the default

currency

A string containing an ISO 4217 currency code. No default value. Full list is availble here NOTE: If this value is not null, number_format will be reassigned to "currency"

unit

A string containing a valid unit identifier from this list Pairs of simple units can be concatenated with "-per-" to make a compound unit. NOTE: If this value is not null, number_format will be reassigned to "unit"

quantile

A number in [0,1], specifying the quantile to use if statistic = "quantile". The default number is 0.5 (i.e. median)

notation

How the number should be displayed. Possible vales are c("standard", "scientific","engineering","compact"). Standard is the default.

width

Optional parameter. Fixed width for widget (in css units), which is passed to htmlwidgets::createWidget.

height

Optional parameter. Fixed height for widget (in css units), which is passed to htmlwidgets::createWidget.

elementId

Optional parameter. String used to define the element containing the widget, which is passed to htmlwidgets::createWidget.