Show a single value in a text span

A summaryTextSpan displays a single value derived from a linked table. Its primary use is with the crosstalk package. Used with crosstalk, a summaryTextSpan displays a value which updates as the data selection changes.

Usage

summaryTextSpan(
  data,
  statistic = c("count", "first", "last", "min", "max", "mode", "concatenate", "unique"),
  column = NULL,
  selection = NULL,
  locale = "navigator.language",
  delim = ", ",
  sort = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)

Arguments

data

Data to summarize, normally an instance of crosstalk::SharedData.

statistic

The aggregation to perform. Possible vales are c("count","first", "last", "min","max","mode", "concatenate", "unique"). first, last and concatenate will select the value based on the index of the column passed to the function. min and max will selected the value based on either d3.ascending (default) or through the function passed to the parameter sort Count is the default.

column

For most statistics, the column of data to summarize. Not used for "count".

selection

Optional parameter. Expression to select a fixed subset of data. May be a logical vector or a one-sided formula that evaluates to a logical vector. If used, the key given to crosstalk::SharedData must be a fixed column (not row numbers).

locale

Optional parameter. For count and unqiue control how the number is displayed. Valid strings include a IETF BCP 47 language tag,

delim

Optional parameter. For concatenate, choose the joining string. Default is ", "

sort

Optional parameter. A sort function used to sort the data to determine the min or max value or the order in which to concatenate the value.

width

Optional parameter. Fixed width for widget (in css units), which is passed to htmlwidgets::createWidget.

height

Optional parameter. Fixed height for widget (in css units), which is passed to htmlwidgets::createWidget.

elementId

Optional parameter. String used to define the element containing the widget, which is passed to htmlwidgets::createWidget.