Create a value box component for a dashboard.

A value box displays a value (usually a number) in large text, with a smaller caption beneath, and a large icon on the right side.

Usage

summaryValueBox(
  data,
  statistic = c("count", "sum", "mean", "pct_total", "sum_pct_total", "min", "max",
    "quantile"),
  column = NULL,
  selection = NULL,
  locale = "navigator.language",
  digits = 0,
  number_format = c("decimal", "percent", "currency", "unit"),
  signDisplay = c("auto", "always", "exceptZero", "negative", "never"),
  currency = NULL,
  unit = NULL,
  quantile = 0.5,
  notation = c("standard", "scientific", "engineering", "compact"),
  caption = NULL,
  icon = NULL,
  color_thresholds = list(domain = c(0.5, 0.8), range = c("red", "gold",
    "mediumseagreen")),
  width = NULL,
  height = NULL,
  elementId = NULL
)

Arguments

data

Data to summarize, normally an instance of crosstalk::SharedData.

statistic

The statistic to compute. Possible vales are c("count", "sum", "mean", "pct_total", "sum_pct_total", "min", "max", "quantile"). Count is the default.

column

For most statistics, the column of data to summarize. Not used for "count" or "pct_total" statistic.

selection

Expression to select a fixed subset of data. May be a logical vector or a one-sided formula that evaluates to a logical vector. If used, the key given to crosstalk::SharedData must be a fixed column (not row numbers).

locale

Control how the number is displayed. Valid strings include a IETF BCP 47 language tag, or the string "navigator.language". Default value is "navigator.language"

digits

Number of decimal places to display, or NULL to display full precision. if number_format == "percent", digits == 0 the statistics is 0.1234, the resulting visualization will show "12%".

number_format

The resulting type of number to display. Options are c("decimal", "percent", "currency","unit"). Decimal is the default.

signDisplay

How to display the sign for the number. Possible values are: c("auto", "always","exceptZero","negative","never"). Auto is the default

currency

A string containing an ISO 4217 currency code. No default value. Full list is availble here NOTE: If this value is not null, number_format will be reassigned to "currency"

unit

A string containing a valid unit identifier from this list Pairs of simple units can be concatenated with "-per-" to make a compound unit. NOTE: If this value is not null, number_format will be reassigned to "unit"

quantile

A number in [0,1], specifying the quantile to use if statistic = "quantile". The default number is 0.5 (i.e. median)

notation

How the number should be displayed. Possible vales are c("standard", "scientific","engineering","compact"). Standard is the default.

caption

Optional parameter. The caption to display beneath the value.

icon

Optional parameter. A string representing the name of a bootstrap icon. See https://icons.getbootstrap.com/ for complete list of names

color_thresholds

a list containing two named vectors, domain and range. The domain is a series of numbers and range is a series of colour strings (hex or html colour names) such that for a value x:

if x < domain[1] then colour == range[1],

if domain[1] <= x < domain[2] then colour == range[2]

...

if domain[n-1] <= x < domain[n] then colour == range[n]

if domain[n] <= x then colour == range[n+1]

As such, the length of the named vector range should be one item longer than the length of domain. The default value for color_threshold is list(domain = c(0.5,0.8), range = c("red","gold","mediumseagreen")). To fix a single colour, let color_thresholds = list(domain = c(-Inf), range=c(unused_color,desired_color))

width

Optional parameter. Fixed width for widget (in css units), which is passed to htmlwidgets::createWidget.

height

Optional parameter. Fixed height for widget (in css units), which is passed to htmlwidgets::createWidget.

elementId

Optional parameter. String used to define the element containing the widget, which is passed to htmlwidgets::createWidget.

Details

See the flexdashboard website for additional documentation: https://pkgs.rstudio.com/flexdashboard/articles/using.html#value-boxes-1

Examples

library(flexdashboard)

valueBox(42, caption = "Errors", icon = "fa-thumbs-down")
#> <span class="value-output" data-caption="Errors" data-icon="fa-thumbs-down" data-color-accent="primary">42</span>
valueBox(107, caption = "Trials", icon = "fa-tag")
#> <span class="value-output" data-caption="Trials" data-icon="fa-tag" data-color-accent="primary">107</span>
valueBox(247, caption = "Connections", icon = "fa-random")
#> <span class="value-output" data-caption="Connections" data-icon="fa-random" data-color-accent="primary">247</span>